Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-keeps-position.window-expected.txt

CONSOLE MESSAGE: line 13: SyntaxError: Parser error

PASS Compiling invalid uncompiled raw handlers should keep the position in event listener list 
